+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Added a CLI block generator
+
 ## v3.3.0 - 2020-11-04
+
 - Reworks Block_Config to remove sections and add method if doing this in the plugin instead
 - Adds Field_Section